Move create_directories_if_not_exist() and get_file_path() to miscellaneous_functions.py
